requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
faiss-cpu==1.12.0
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.2
numpy==2.3.3
openai==1.108.0
packaging==25.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
